fix shellcheck

+++ latest/overlay/etc/owncloud.d/40-apps.sh
@@ -1,33 +1,35 @@
 #!/usr/bin/env bash
 
 if dpkg --compare-versions "$(occ config:system:get version | tail -1)" "ge" "10.0"; then
-  for VAL in $(echo "${OWNCLOUD_APPS_INSTALL}" | tr "," " "); do
-    if [[ "${VAL}" =~ ^http.* ]]; then
-      NAME=$(basename "${VAL}" | cut -d? -f1)
+  if [[ -n "${OWNCLOUD_APPS_INSTALL}" ]]; then
+    for VAL in $(echo "${OWNCLOUD_APPS_INSTALL}" | tr "," " "); do
+      if [[ "${VAL}" =~ ^http.* ]]; then
+        NAME=$(basename "${VAL}" | cut -d? -f1)
 
-      echo "Downloading ${NAME} app..."
-      curl -sSfLo "/tmp/${NAME} "-H "Accept: application/octet-stream" "${VAL}"
+        echo "Downloading ${NAME} app..."
+        curl -sSfLo "/tmp/${NAME} "-H "Accept: application/octet-stream" "${VAL}"
 
-      echo "Installing ${NAME} app..."
-      occ market:install -n -l "/tmp/${NAME}"
+        echo "Installing ${NAME} app..."
+        occ market:install -n -l "/tmp/${NAME}"
 
-      if [[ $OWNCLOUD_APPS_INSTALL_MAJOR == "true" ]]; then
-        occ market:upgrade -n -q --major -l "/tmp/${NAME}"
+        if [[ $OWNCLOUD_APPS_INSTALL_MAJOR == "true" ]]; then
+          occ market:upgrade -n -q --major -l "/tmp/${NAME}"
+        fi
+
+        echo "Deleting ${NAME} tarball..."
+        rm -f "/tmp/${NAME}"
+      else
+        if [[ ! -d ${OWNCLOUD_VOLUME_APPS}/${VAL} ]]; then
+          echo "Installing ${VAL} app..."
+          occ market:install -n "${VAL}"
+        fi
+
+        if [[ $OWNCLOUD_APPS_INSTALL_MAJOR == "true" ]]; then
+          occ market:upgrade -n -q --major "${VAL}"
+        fi
       fi
-
-      echo "Deleting ${NAME} tarball..."
-      rm -f "/tmp/${NAME}"
-    else
-      if [[ ! -d ${OWNCLOUD_VOLUME_APPS}/${VAL} ]]; then
-        echo "Installing ${VAL} app..."
-        occ market:install -n "${VAL}"
-      fi
-
-      if [[ $OWNCLOUD_APPS_INSTALL_MAJOR == "true" ]]; then
-        occ market:upgrade -n -q --major "${VAL}"
-      fi
-    fi
-  done
+    done
+  fi
 
   if [[ -n "${OWNCLOUD_APPS_ENABLE}" ]]; then
     for VAL in $(echo "${OWNCLOUD_APPS_ENABLE}" | tr "," " "); do
